organized file structure, added example tests, fixed Seek return
